requirements.txt
requests
redis
python-dotenv
PyQt5